requirements.txt
requests==2.32.3
PyGithub==2.3.0
python-dotenv==1.0.1
